Use workaround filename as schema key name FOLIO-573
Removing the symlink workaround.
This is the bare minimum to address FOLIO-573:
When any schema file refers to an additional schema file, then use the
filename of that referenced second schema as the "key" name in the
RAML "schemas" section, and wherever that schema is utilised in RAML
and schema files.